First pass move files.
