+++ libshortfin/src/CMakeLists.txt
@@ -18,7 +18,8 @@ shortfin_public_library(
     shortfin
   COMPONENTS
     shortfin_array
-    shortfin_local_system
+    shortfin_local
+    shortfin_process
     shortfin_support
     shortfin_systems_amdgpu
     shortfin_systems_host
