+++ libshortfin/src/shortfin/local/CMakeLists.txt
@@ -6,17 +6,21 @@
 
 shortfin_cc_component(
   NAME
-    shortfin_array
-  HDRS
-    array.h
-    api.h
-    dtype.h
-    storage.h
-  SRCS
-    array.cc
-    dtype.cc
-    storage.cc
-  COMPONENTS
     shortfin_local
+  HDRS
+    device.h
+    scope.h
+    system.h
+  SRCS
+    device.cc
+    scope.cc
+    system.cc
+  COMPONENTS
+    shortfin_process
     shortfin_support
+  DEPS
+    iree_base_base
+    iree_hal_hal
+    iree_modules_hal_hal
+    iree_vm_vm
 )
