+++ libshortfin/src/shortfin/process/CMakeLists.txt
@@ -4,8 +4,16 @@
 # https://llvm.org/LICENSE.txt for license information. SPDX-License-Identifier:
 # Apache-2.0 WITH LLVM-exception
 
-add_subdirectory(array)
-add_subdirectory(local)
-add_subdirectory(process)
-add_subdirectory(support)
-add_subdirectory(systems)
+shortfin_cc_component(
+  NAME
+    shortfin_process
+  HDRS
+    worker.h
+  SRCS
+    worker.cc
+  COMPONENTS
+    shortfin_support
+  DEPS
+    iree_base_base
+    iree_base_loop_sync
+)
